telegram_bot/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-telegram-bot==21.0.1
requests==2.32.3
pydantic==2.7.1
python-dotenv==1.0.1